requirements.txt
flet
python-i18n
pydantic
pyinstaller
beanie